--removed useless code, moved python inference file to xconsortia, and renamed the helper files
