+++ src/gtex/downloader.js
@@ -158,6 +158,6 @@ export class Downloader {
       throw new Error(`Dataset has fewer than ${minCount} cell. Cell count: ${dataset.dataset_cell_count}`);
     }
 
-    dataset.rna_source = 'nucleus';
+    dataset.dataset_rna_source = 'nucleus';
   }
 }
