+++ src/util/fs.js
@@ -1,8 +1,10 @@
 import { access, constants, mkdir, open, writeFile } from 'node:fs/promises';
-import { install, Agent } from 'undici';
+//import { install, Agent } from 'undici';
 import { concurrentMap } from './concurrent-map.js';
+import undici from 'undici';
 
-install();
+const { install, Agent } = undici;
+install?.();
 
 /**
  * Ensure that all specified directories are created
